refactor: move Filtered*State and Wrapped*State to swirlds-state-api
