+++ hedera-node/hedera-app-spi/src/main/java/module-info.java
@@ -20,6 +20,5 @@ module com.hedera.node.app.spi {
     exports com.hedera.node.app.spi.validation;
     exports com.hedera.node.app.spi.workflows.record;
     exports com.hedera.node.app.spi.authorization;
-    exports com.hedera.node.app.spi.state;
     exports com.hedera.node.app.spi.metrics;
 }
