+++ hedera-node/hedera-app-spi/src/test/java/com/hedera/node/app/spi/state/SchemaTest.java
@@ -20,9 +20,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.hedera.hapi.node.base.SemanticVersion;
-import com.hedera.node.app.spi.fixtures.state.MapWritableStates;
 import com.swirlds.platform.test.fixtures.state.TestSchema;
 import com.swirlds.state.lifecycle.Schema;
+import com.swirlds.state.test.fixtures.MapWritableStates;
 import com.swirlds.state.test.fixtures.StateTestBase;
 import java.util.ArrayList;
 import java.util.Collections;
