+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/services/MigrationContextImpl.java
@@ -20,11 +20,11 @@ import static java.util.Objects.requireNonNull;
 
 import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.node.app.ids.WritableEntityIdStore;
-import com.hedera.node.app.spi.state.FilteredWritableStates;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.platform.state.MerkleStateRoot;
 import com.swirlds.state.lifecycle.MigrationContext;
 import com.swirlds.state.lifecycle.info.NetworkInfo;
+import com.swirlds.state.spi.FilteredWritableStates;
 import com.swirlds.state.spi.ReadableStates;
 import com.swirlds.state.spi.WritableStates;
 import edu.umd.cs.findbugs.annotations.NonNull;
