+++ platform-sdk/swirlds-state-api/src/main/java/com/swirlds/state/spi/FilteredReadableStates.java
@@ -14,12 +14,8 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.state;
+package com.swirlds.state.spi;
 
-import com.swirlds.state.spi.ReadableKVState;
-import com.swirlds.state.spi.ReadableQueueState;
-import com.swirlds.state.spi.ReadableSingletonState;
-import com.swirlds.state.spi.ReadableStates;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 import java.util.Set;
