+++ platform-sdk/swirlds-state-api/src/main/java/com/swirlds/state/spi/FilteredWritableStates.java
@@ -14,12 +14,8 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.state;
+package com.swirlds.state.spi;
 
-import com.swirlds.state.spi.WritableKVState;
-import com.swirlds.state.spi.WritableQueueState;
-import com.swirlds.state.spi.WritableSingletonState;
-import com.swirlds.state.spi.WritableStates;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 import java.util.Set;
