+++ platform-sdk/swirlds-state-api/src/main/java/com/swirlds/state/spi/WrappedWritableSingletonState.java
@@ -14,10 +14,8 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.state;
+package com.swirlds.state.spi;
 
-import com.swirlds.state.spi.WritableSingletonState;
-import com.swirlds.state.spi.WritableSingletonStateBase;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 /**
