+++ platform-sdk/swirlds-state-api/src/test/java/com/swirlds/state/spi/EmptyWritableStatesTest.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.state;
+package com.swirlds.state.spi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import com.swirlds.state.spi.EmptyWritableStates;
 import com.swirlds.state.test.fixtures.StateTestBase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
