+++ platform-sdk/swirlds-state-api/src/test/java/com/swirlds/state/spi/WrappedWritableKVStateTest.java
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.state;
+package com.swirlds.state.spi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.anyString;
 
-import com.swirlds.state.spi.WritableKVStateBase;
 import com.swirlds.state.test.fixtures.MapWritableKVState;
 import com.swirlds.state.test.fixtures.StateTestBase;
 import java.util.HashMap;
