+++ platform-sdk/swirlds-state-api/src/testFixtures/java/com/swirlds/state/test/fixtures/MapWritableStates.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi.fixtures.state;
+package com.swirlds.state.test.fixtures;
 
 import static java.util.Objects.requireNonNull;
 
@@ -26,8 +26,6 @@ import com.swirlds.state.spi.WritableQueueStateBase;
 import com.swirlds.state.spi.WritableSingletonState;
 import com.swirlds.state.spi.WritableSingletonStateBase;
 import com.swirlds.state.spi.WritableStates;
-import com.swirlds.state.test.fixtures.MapReadableStates;
-import com.swirlds.state.test.fixtures.MapWritableKVState;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.Collections;
